requirements.txt
langchain-text-splitters
langchain 
langchain-core 
langchain-community
pinecone
ipykernel
pypdf
google-generativeai
Flask